Clean up chain reader, removed map decoder, as it will no longer be used (#242)
