+++ pkg/types/codec.go
@@ -30,7 +30,7 @@ type Codec interface {
 }
 
 type TypeProvider interface {
-	CreateType(itemType string, forceSlice, forEncoding bool) (any, error)
+	CreateType(itemType string, forEncoding bool) (any, error)
 }
 
 type RemoteCodec interface {
